+++ src/org/andreschnabel/jprojectinspector/metrics/OfflineMetric.java
@@ -7,6 +7,6 @@ public interface OfflineMetric {
 	public String getName();
 	public String getDescription();
 
-	public float measure(File repoRoot) throws Exception;
+	public double measure(File repoRoot) throws Exception;
 
 }
