+++ src/org/andreschnabel/jprojectinspector/metrics/javaspecific/JavaTestFrameworkDetector.java
@@ -55,7 +55,7 @@ public class JavaTestFrameworkDetector implements OfflineMetric {
 	}
 
 	@Override
-	public float measure(File repoRoot) throws Exception {
+	public double measure(File repoRoot) throws Exception {
 		return checkForFramework(repoRoot);
 	}
 }
