+++ src/org/andreschnabel/jprojectinspector/metrics/project/CodeFrequency.java
@@ -39,7 +39,7 @@ public class CodeFrequency implements OnlineMetric {
 	}
 
 	@Override
-	public float measure(Project p) throws Exception {
+	public double measure(Project p) throws Exception {
 		return countCodeFrequencyForProj(p);
 	}
 }
