+++ src/org/andreschnabel/jprojectinspector/metrics/project/ContributorsTestCommit.java
@@ -25,7 +25,7 @@ public class ContributorsTestCommit implements OfflineMetric {
 	}
 
 	@Override
-	public float measure(File repoRoot) throws Exception {
+	public double measure(File repoRoot) throws Exception {
 		return countNumTestContributors(repoRoot);
 	}
 }
