+++ src/org/andreschnabel/jprojectinspector/tests/TestCommon.java
@@ -11,5 +11,5 @@ public final class TestCommon {
 	public static final String TEST_SRC_TEST_FILENAME = "testdata"+File.separator+"PointsTest.java";
 	public static final Project THIS_PROJECT = new Project("0x17", "JProjectInspector");
 
-	public static final float DELTA = 0.00001f;
+	public static final double DELTA = 0.00001f;
 }
