+++ src/org/andreschnabel/jprojectinspector/tests/offline/metrics/RoughFunctionCoverageTest.java
@@ -10,10 +10,10 @@ import java.util.Map;
 public class RoughFunctionCoverageTest {
 	@Test
 	public void testApproxFunctionCoverage() throws Exception {
-		Map<String, Float> coverages = RoughFunctionCoverage.approxFunctionCoverage(new File("testdata"));
+		Map<String, Double> coverages = RoughFunctionCoverage.approxFunctionCoverage(new File("testdata"));
 
 		for(String lang : coverages.keySet()) {
-			float cov = coverages.get(lang);
+			Double cov = coverages.get(lang);
 			Assert.assertTrue(cov >= 0.0f);
 			Assert.assertTrue(cov <= 1.0f);
 		}
